% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkbcgeneral-package.R
\docType{package}
\name{nkbcgeneral-package}
\alias{nkbcgeneral}
\alias{nkbcgeneral-package}
\title{nkbcgeneral: Generella verktyg för NKBC}
\description{
Generella verktyg för bearbetning av data från Nationellt
    kvalitietsregister för bröstcancer (NKBC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cancercentrum.bitbucket.io/nkbcgeneral/}
  \item \url{http://bitbucket.org/cancercentrum/nkbcgeneral}
  \item Report bugs at \url{https://bitbucket.org/cancercentrum/nkbcgeneral/issues}
}

}
\author{
\strong{Maintainer}: John Lövrot \email{john.lovrot@sll.se} (\href{https://orcid.org/0000-0002-9339-8059}{ORCID})

Authors:
\itemize{
  \item Lina Benson (\href{https://orcid.org/0000-0002-5393-0055}{ORCID})
}

Other contributors:
\itemize{
  \item RCC [copyright holder, funder]
}

}
\keyword{internal}
